% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrichr_terms}
\alias{enrichr_terms}
\title{Enrichr gene sets}
\format{A tibble with 123 rows and 10 columns
\describe{
  \item{library}{the library or source the gene sets where derived from}
  \item{year}{release year of the library}
  \item{n_terms}{number of terms (gene sets)}
  \item{gene_coverage}{number of unique genes across all terms}
  \item{n_genes_per_term}{average number of genes per term}
  \item{category}{broad category for the gene set library}
  \item{description}{description of the membership in the gene set}
  \item{data}{list column containing the symbols for each term}
}}
\source{
\url{http://amp.pharm.mssm.edu/Enrichr/}
}
\usage{
enrichr_terms
}
\description{
A dataset containing all gene sets from the Enrichr database including
metadata. Release 2019-01-23.
}
\details{
Where multiple versions are available, only the most recent release
is included.

Please acknowledge Enrichr in your publications by citing the following references:

\cite{Chen EY, Tan CM, Kou Y, Duan Q, Wang Z, Meirelles GV, Clark NR, Ma'ayan A. Enrichr: interactive and collaborative HTML5 gene list enrichment analysis tool. BMC Bioinformatics. 2013;128(14)}

\cite{Kuleshov MV, Jones MR, Rouillard AD, Fernandez NF, Duan Q, Wang Z, Koplev S, Jenkins SL, Jagodnik KM, Lachmann A, McDermott MG, Monteiro CD, Gundersen GW, Ma'ayan A. Enrichr: a comprehensive gene set enrichment analysis web server 2016 update. Nucleic Acids Research. 2016; gkw377}
}
\keyword{datasets}
